Change heuristic-based scheduling to a 2-phase one for exchange client
Exchange client schedules HTTP buffer clients to fetch pages based on
heuristics. This works under the assumption that the data transfering is
under a rather stable and evenly distributed way for all the sinks.
There could be cases where the coming pages are way larger than the
previous ones. This can cause a sudden increase in the buffer size. In
the worst case, if the cluster is large enough (e.g. > 300) and has
large number of concurrent running queries (e.g. > 40), this can lead
to a worst case of 300 X 40 X 16MB = 192GB memory usage. This patch
splits the fetching protocol into a two-phase manner by first retrieving
the page sizes, reversing memory on the source, then fetching the pages
within the memory limit.