Introduce empty block
When a cell is empty, an block instance will be created with position
count equal to zero. This is relatively common in production especially
when tables have empty maps or arrays. Creating blocks with zero
positions can cause GC pressure. This patch introduces EmptyBlock as a
singleton to replace such empty instances.