Use big arrays for ReferenceCountMap
ReferenceCountMap maps from objects to reference counts. Its keys are
stored in a single object array. In some extreme cases, the array grows
to hundreds of millions of objects. This can cause G1GC to push huge
amounts of object pointers to a task queue, which can both lead to GC
overhead as well as queue overflow.