Fix buffer client acknowledge logic
(1) A client can received empty pages. In such case, there is no need
to acknowledge the pages. This helps reducing traffic.
(2) The acknowledged token should be the next one instead of current
one, so that the server can remove all pages that have been received.
Acknowledging the current token does not help removing the received
pages on the server side because they should have been removed in a
previous request.