TODO: Raghav: Finish query on dispatcher when it is finished on coordinator
This is just a temp change; we shouldn't let coordinator sends ack back
to the dispatcher.

This aligns with how we should fetch stats from coordinator.
A query on dispatcher that is in "submitted" state needs to finish once
the real query on coordinator finishes. Or if the coordinator is not
reachable (e.g., crashed). We need to properly handle the finishing of
those submitted queries.